feat: add CustomerDeliveryDropdownCard
